refactor: refactored the location screen
